Create The_Code.py
Talos the mighty! Talos the unerring! Talos the unassailable! To you we give praise!
We are but maggots, writhing in the filth of our own corruption!
 While you have ascended from the dung of mortality, and now walk among the stars!
But you were once man! Aye! 
And as man, you said, "Let me show you the power of Talos Stormcrown, born of the North, where my breath is long winter. I breathe now, in royalty, and reshape this land which is mine. I do this for you, Red Legions, for I love you."
Aye, love. Love! Even as man, great Talos cherished us.
 For he saw in us, in each of us, the future of Skyrim! The future of Tamriel!